requirements.txt
streamlit
transformers
torch
torchvision
accelerate
bitsandbytes
langchain
langchain-community
sentence-transformers
chromadb
PyMuPDF
pillow